[ty] Move match pattern types out of reachability
